chore: rename app-loader to bootstrapper to avoid confusion
