+++ apps/trading/components/bootstrapper/bootstrapper.tsx
